+++ fendermint/vm/interpreter/src/chain.rs
@@ -3,7 +3,6 @@
 use crate::fvm::state::ipc::GatewayCaller;
 use crate::fvm::store::ReadOnlyBlockstore;
 use crate::fvm::{topdown, EndBlockOutput, FvmApplyRet};
-use crate::selector::{GasLimitSelector, MessageSelector};
 use crate::{
     fvm::state::FvmExecState,
     fvm::FvmMessage,
@@ -17,6 +16,7 @@ use fendermint_tracing::emit;
 use fendermint_vm_actor_interface::ipc;
 use fendermint_vm_event::ParentFinalityMissingQuorum;
 use fendermint_vm_message::ipc::ParentFinality;
+use fendermint_vm_message::signed::SignedMessage;
 use fendermint_vm_message::{
     chain::ChainMessage,
     ipc::{BottomUpCheckpoint, CertifiedMessage, IpcMessage, SignedRelayedMessage},
@@ -117,7 +117,7 @@ where
         (chain_env, state): Self::State,
         mut msgs: Vec<Self::Message>,
     ) -> anyhow::Result<Vec<Self::Message>> {
-        msgs = messages_selection(msgs, &state)?;
+        msgs = select_messages(msgs, &state)?;
 
         // Collect resolved CIDs ready to be proposed from the pool.
         let ckpts = atomically(|| chain_env.checkpoint_pool.collect_resolved()).await;
@@ -232,7 +232,7 @@ where
                 ChainMessage::Signed(signed) => {
                     // We do not need to check against the minimium base fee because the gas market
                     // is guaranteed to be capped at it anyway.
-                    if signed.message.gas_fee_cap < gas_market.base_fee {
+                    if !signed.includable(&gas_market.base_fee) {
                         // We do not accept blocks containing transactions with gas parameters below the current base fee.
                         // Producing an invalid block like this should penalize the validator going forward.
                         return Ok(false);
@@ -582,16 +582,16 @@ fn relayed_bottom_up_ckpt_to_fvm(
     Ok(msg)
 }
 
-/// Selects messages to be executed. Currently, this is a static function whose main purpose is to
-/// coordinate various selectors. However, it does not have formal semantics for doing so, e.g.
-/// do we daisy-chain selectors, do we parallelize, how do we treat rejections and acceptances?
-/// It hasn't been well thought out yet. When we refactor the whole *Interpreter stack, we will
-/// revisit this and make the selection function properly pluggable.
-fn messages_selection<DB: Blockstore + Clone + 'static>(
+/// Selects messages to be included in a proposal.
+/// Currently, this is a static function whose main purpose is to compose various hard-coded selectors.
+/// When we refactor the whole *Interpreter stack, we will revisit this and make the selection
+/// function properly pluggable.
+fn select_messages<DB: Blockstore + Clone + 'static>(
     msgs: Vec<ChainMessage>,
     state: &FvmExecState<DB>,
 ) -> anyhow::Result<Vec<ChainMessage>> {
-    let mut user_msgs = msgs
+    // Ensure we only have signed messages.
+    let signed = msgs
         .into_iter()
         .map(|msg| match msg {
             ChainMessage::Signed(inner) => Ok(inner),
@@ -599,13 +599,27 @@ fn messages_selection<DB: Blockstore + Clone + 'static>(
         })
         .collect::<anyhow::Result<Vec<_>>>()?;
 
-    // Currently only one selector, we can potentially extend to more selectors
-    // This selector enforces that the total cumulative gas limit of all messages is less than the
-    // currently active block gas limit.
-    let selectors = vec![GasLimitSelector {}];
-    for s in selectors {
-        user_msgs = s.select_messages(state, user_msgs)
-    }
+    let total_gas_limit = state.block_gas_tracker().available();
+    let gas_market = state.block_gas_tracker().current_gas_market();
+    let mut total_gas_limit_consumed = 0;
 
-    Ok(user_msgs.into_iter().map(ChainMessage::Signed).collect())
+    // Selector functions to compose.
+    let is_includable = |msg: &SignedMessage| msg.includable(&gas_market.base_fee);
+    let cumulative_gas_limit = |msg: &SignedMessage| {
+        let gas_limit = msg.message.gas_limit;
+        let accepted = total_gas_limit_consumed + gas_limit <= total_gas_limit;
+        if accepted {
+            total_gas_limit_consumed += gas_limit;
+        }
+        accepted
+    };
+
+    // Perform the selection.
+    let selected = signed
+        .into_iter()
+        .filter(is_includable)
+        .take_while(cumulative_gas_limit)
+        .map(ChainMessage::Signed)
+        .collect::<Vec<_>>();
+    Ok(selected)
 }
